+++ client/src/components/FormInput.js
@@ -1,0 +1,32 @@
+
+import "./formInput.css";
+import {useState} from "react";
+const FormInput = (props) => {
+
+    const [focused, setFocused] = useState(false);
+    const {label, errorMessage, placeholder, onChange, id, ...inputProps} = props;
+
+    const handleFocus = (e) => {
+        setFocused(true);
+    }
+
+    return (
+        <div className = "formInput">
+            <label>{label}</label>
+            <input {...inputProps} 
+                placeholder = {placeholder}
+                onChange={onChange}
+                onBlur = {handleFocus}
+                //onFocus = {() => inputProps.name==="confirmPassword" && setFocused(true)}
+                focused = {focused.toString()}
+            />
+            <span>{errorMessage}</span>
+        </div> 
+    )
+}
+
+export default FormInput;
+
+
+//blur - click and leave
+//focus - one click
